+++ src/App.js
@@ -6,7 +6,7 @@ import { Main } from './components/main/Main';
 import { Footer } from './components/footer/Footer';
 import { ProductListing } from '../src/components/main/product_listing/ProductListing';
 import { CartContainer } from '../src/components/cart/CartContainer/CartContainer';
-import { ConfirmOrderContainer } from './components/cart/CartContainer/BuyForm/ConfirmOrderContainer/ConfirmOrderContainer';
+import { ConfirmOrderContainer } from './components/cart/CartContainer/BuyBox/ConfirmOrderContainer/ConfirmOrderContainer';
 import { useResponse } from './hooks/useResponse';
 
 function App() {
