+++ src/components/cart/CartContainer/BuyBox/ConfirmOrderContainer/OrderForm/Modal/Modal.module.css
@@ -1,5 +1,5 @@
 .modal_layout {
-	position: absolute;
+	position: fixed;
 	top: 0;
 	left: 0;
 	height: 100vh;
